fix: handle special chars in window title.
Some window titles can't be passed to `find_titled()`.  We could probably escape
them appropriately, but it's easier (and more robust) just to get all the
windows which include a suitable (safe) part of the title and then find the one
whose string matches.

Since the window title is the resolved path to the file in question, and we
don't use `/`, it's guaranteed to be at least one node deep, so we just get the
outer dir and use that.